+++ test/spec/SpecHelper.hs
@@ -162,6 +162,7 @@ baseCfg = let secret = encodeUtf8 "reallyreallyreallyreallyverysafe" in
   , configInternalSCLoadSleep       = Nothing
   , configInternalSCRelLoadSleep    = Nothing
   , configServerTimingEnabled       = True
+  , configServerOtelEnabled         = False
   }
 
 testCfg :: AppConfig